updated 'formatInPounds' method within 'Numbers' class and added a timer-based counting class
The 'formatInPounds' method of the 'Numbers' class is now more robust
and allows users to specify the number of zeros that should initially
appear in place of pounds (when there are less pound decimals than this
number).

The 'CountTimer' class will count either up or down from a start index
to a target index. It dispatches update, complete and total complete
events/signals and can be set to loop.